requirements.txt
pytorch_pretrained_bert
streamlit==0.82.0
numpy==1.17.2
transformers==2.10.0
newspaper3k==0.2.8
nltk==3.5
pdfplumber
docx2txt
